Fixed to share ndt struct
